consolidate value-conversion code
